CI reuse environ setup between tests
